qcacmn: Add connection state check before clearing mlo vdev
Currently the driver is trying to notify the userspace
upon disconnect on partner link. If initial connection on
partner link fails and a NB disconnect was triggered at the
same time, moving the state machine to disconnecting, the driver
is clearing the mlo vdev flags.

To address this issue, check for the connection state
and clean up the mlo vdev flags after disconnect is complete.

CRs-Fixed: 3309091
Change-Id: I2e52ae225d33f3ff7be06e0a3abe8709f42012ee