qcacmn: Fix dp_main.c documentation
The kernel-doc script identified a large number of documentation
issues in dp/wifi3.0/dp_main.c, so fix those issues. In addition,
there are a number of instances where public functions have their
implementation documented instead of having their interface
documented, so remove that documentation. The documentation at the
interface will be added separately if it is not already present.

Change-Id: Id9bcfd37d690b72ec7a04f25d1b91b0210bef8e3
CRs-Fixed: 3373154