qcacmn: Consider AFC response with 0 freq/channel objects as valid
With the current behavior, a NULL AFC response is treated to be invalid
and is not processed. Hence the AP continues to operate in its current
state. To resolve this, the NULL AFC response should be treated as if
there are no SP channels available in that geo-location and should be
processed as follows:
A) If there are no Tx channels available(in all power modes) then the AP
   must be brought down.
B) Else AP must choose the best channel in the best power mode.

To implement above cases, remove all the checks in the driver that tries
to block the NULL response and finally call upper layer functions
(hostapd/acs) so that they can decide whether to bring down the AP or
choose a channel from the available power modes. Modify
'reg_set_ap_pwr_and_update_chan_list' to bypass setting power type when
ap_pwr_type is REG_CURRENT_MAX_AP_TYPE. This allows the current channel
list to be updated and the upper layers to be informed about this update.

Change-Id: Id048fa9ee402d537f352e3caa2b53729518ea37e
CRs-Fixed: 3565567