qcacmn: Keep correct set of 6 GHz channels enabled in super channel list
Currently if fcc_constraint 0/2 is set and there are
active 6 GHz STA/P2P client connections, host keeps the
bonded channels enabled, but in super channel list
incorrect set of frequencies remain active. This
happens because the active frequency start and end
indexes are not according to the super channel list
indexes.

To address this issue, convert the active 6 GHz
range start and end indexes according to the super
channel list indexes.

CRs-Fixed: 3595197
Change-Id: I8bb3a5093eb3b8bd970fcd371a4939570207385b