qcacmn: Check for target_suspend state before issuing WMI
In OPT_WIFI_DP we perform a system_wakeup before issuing WMI command
to FW to disable low power mode. But as this is not a blocking call,
WMI command could be sent to FW before target is completely
resumed. This will trigger an intentional safety assert.
This change will wait 500ms for target to resume and send
the WMI command after ensuring target is not in suspend state.

Change-Id: I380bd9fd4e292caba5f32f80e65f3c5c992dae3c
CRs-Fixed: 3604250