qcacmn: Try SLO if 2 link MLO connection fails
Currently, max retail for connection is set to 1
it means only 2 trials are allowed, in case of 2
link MLO AP logic works fine but for n link only
first 2 combinations will be tried and SLO will
not be tried as no of retries will exceed than max
trials.

Fix is to update max retries to num_of_mlo_link - 1,
so that all combination tll SLO can be tried.

Change-Id: Id3e1da733d77eae79f2389b98ee9a664e44e6a2c
CRs-Fixed: 3607628