qcacmn: Add new API to get peer info from DP
Add new API to get some information of peer from DP
layer, like which vdev id does peer under, whether
is a mld device or not, and what is the peer state.
These information will be used for CLD DP to
send unicast packet on correct dp link.

Note:
If it is legacy client, vdev id is the one that peer under.
If it is MLO STA, vdev id is the primary(assoc) link that peer under.

Change-Id: Iea34e08baae01e8a7cc49e94ea0b49c5b18c0fe6
CRs-Fixed: 3679384