qcacmn: Increase soc wake time for ipci devices
Currently for IPCI devices, SoC wake request and release messages
are serialized, i.e., a new request cannot make it to the firmware
unless the current request is completed. In some cases, when system
is heavily loaded, SoC wake request cannot make it to the firmware
due to a previous SoC wake release resulting in a timeout.

Increase the SoC wake timeout to 1 second to avoid this. Since this
is a worst case value, increasing it should be fine.

Change-Id: Ic265c0076233378bc6db7b2a99f7a0b7d7a4b2bc
CRs-Fixed: 3748244