qcacmn: Support setting mcast vdev flag to dp side
For MLO SAP case, whatever it is single link or
multi-link, we expect that it can always follow
up with global sequence number for broadcast/
multicast packet.

Then it needs the mcast vdev flag set in the DP side
to distinguish if it is mlo sap or legacy sap,
and also dp side will use this vdev as the primary
vdev doing data forwarding.
So add the logic to set this flag if the first
mlo sap link to be added.

Change-Id: I55a22fe224fc0044c072443915302e8e75399bfb
CRs-Fixed: 3750541