qcacmn: Add optional argument to treat NOL as enabled
The regulatory API 'reg_get_band_from_cur_chan_list' is used
for copying the channels which are enabled from the input
channel list based on the input band flags. Currently this
API treats NOL as disabled and excludes the NOL channels
from the resultant channel list.

Make this configurable by adding an additional optional
argument to pass a flag to determine whether to treat NOL
channels are enabled or disabled.

Also, add the dispatcher level changes for the callers to
pass this optional argument from higher layers.

Change-Id: I49e855a6c86dcc8dfa731d1e7599db242f786c0e
CRs-Fixed: 3703106