qcacmn: Change prototype of Agile DFS config API
Currently the API used for setting the preCAC config takes the input
config flag as an unsigned 32 bit integer. Since the flag is used to
check only if the config is enabled or disabled, use boolean as the
datatype instead of unsigned 32 bit integer.

Make changes in necessary function header files as well.

Change-Id: Ib71bb01907a1d93862c388d767fc9301046a5a01
CRs-Fixed: 3751530