qcacmn: fix initialization of filter event and tx completion path
This change is to create new filter event for opt_dp_ctrl,
initialize the event before ipa connect pipes
and fix tx completion path for opt_dp_ctrl.

Need to make new event for filter add/delete in opt_dp_ctrl
cause filter addition/deletion of both the features
opt_dp and opt_dp_ctrl can be triggered at same time.

It is required to move the case TX_STATUS_OK below
TX_STATUS_DROP to avoid fallthrough, as we need to
unvote the clock in case of ctrl pkt drop in fw and
should not be done in the case of success.

Change-Id: I6b953fe497ea64ca0481f8af803bd8c6d226e007
CRs-Fixed: 3776854