qcacmn: Update score of pending candidates based on prev failures
Currently, BSSIDs which are part of avoid list are not considered
while updating the candidate list due to which previous calculated
score of deny list entries are getting updated with new score. So,
the same MLO entry is considered and connection is tries again.

Fix is not to considered avoided links for which API is introduced
to check if any dlm candidate is prsent in the list then give min
score to that candidate and add at last. Print the candidate list
only if new entry is found or any dlm entry is present.

Change-Id: I430ccb4b28a86dae4714c636a44bf789e17d977f
CRs-Fixed: 3791454