qcacmn: Support to indicate WDS STA to associated APs operating BW
In mixed bandwidth scenarios, there is a need for the station (STA)
to be aware of the access point’s (AP) bandwidth. For example, if
the AP operates at 80 MHz, STA1 also operates at 80 MHz, and STA2
operates at 20 MHz, STA2 should be informed about the AP’s
bandwidth (80 MHz).

To satisfy the above stated requirement, a new WMI command has been
introduced. This command provides an indication of the current
operating bandwidth of the associated access point (AP) to the WDS
STA's FW. The WDS station sends this command to firmware after
establishing a peer association and whenever there is a dynamic
change in the operating mode reported by the associated AP.

Change-Id: Ib5946819ea4e506f75b52c4064b5a09bb8ccf984
CRs-Fixed: 3849509