qcacmn: Parse INIs without depending on delimiter
Currently while parsing and validating the ini file, if there is
no END delimiter at the end of the ini file, host treats the file
as invalid and doesn't apply the ini configurations. This leads to
a scenario where all INIs are configured with default values.
To avoid it, while validating the ini key-value pair, return the
file as invalid when only unpermitted characters are present.

Change-Id: I110dcbb490f4c64a831e51284d689f68227ba1b0
CRs-Fixed: 3812101