qcacmn: Change Maximum Length of TPE IE for 11BE
TPE IE signaling was enhanced as part of IEEE P802.11be/D4.1.
to add support for sending Tx powers for 11BE - when the
operating bandwidth is 320 MHz or if the channel is punctured.

Accordingly, the maximum length of the TPE was increased to 18.
But when we process the TPE IEs in Channel switch Wrapper IE in
beacon frames, the max TPE IE length is set as 9 (IEEE802.11AX).
Thus, we end up discarding the beacon and action frames.

This causes failure in seamless channel change and thus the STA
vap disconnect, scan and connect again.

To fix this, change the max length to 18 based on the
IEEE802.11BE/D4.1.
For 320 MHz, the maximum number of subchannels is 16.
And 2 extra octets for (1) Transmit Power Information Field and
(2) Extension Transmit PSD Information Field

Also, to ensure forward compatibility, maximum length check is
not needed for variable length IEs.

Change-Id: Ied5ea0a3c82977eb36204f2f7518437f2d52abb8
CRs-Fixed: 3820449