qcacmn: Link PTQM error handling
Peer migration framework can notify the status synchronously
in below cases

- Request fails before giving the request to the core PTQM
  framework. In this case, error is returned to the caller.
- Core framework returns failure if it is not able to assign
  a new primary. In this case, error is notified to the caller
  via a callback.

The callback releases link request context when no peer requests
are pending. This can lead to a use after free scenario if the
caller accesses link request context after the callback is called
synchronously for the last peer request.

To fix this, the index of the link request context in MLD device
is used to check if the link request context is released
by the callback before processing the next peer request.

Change-Id: I12ac4de8e3ec5b0e1ab1a20aaf4892ceba918bdc
CRs-Fixed: 3830027