qcacmn: Fix reap logic in DP Tx Comp handler
In  DP Tx comp handler while reaping the desc,
the condition for more desc available in ring
always returned zero as the ring hp was cached,
as a result there where no desc found in the
ring to process even though the quota was
available.
1) Fix the API by using the non-cached hp
while checking for more desc in the ring.
2) Add conditions to make sure total desc
process should not exceed napi quota.

Change-Id: I3e4503fc36c133c14485af3cee2a9fdae74222e8
CRs-Fixed: 3839312