qcacmn: Fix nbuf freeing before unmap in opt_dp_ctrl
This change is to fix the case where nbuf is getting free without
unmap. This is observed exceptional path where work
queue schedule for ctrl pkt fails because list used for storing
buffers reached max capacity. It has been seen that we are
unmapping the buffer after calling API for work schedule causing this
issue in expectional path.
In normal scenario we schedule WQ for ctrl pkt and then unmap the nbuf
in later stage and it doesn't hit this issue because buffers are always
unmapped immediately after enqueue to work queue. But in this case,
we are hitting the scenario where nbuf freed before unmapping as
it is in same context.

To resolve this unmapping the buffer before calling API for work
scheduling.

CRs-Fixed: 3844894
Change-Id: I86504d057477b87e7b0b58ff433bb358ee20de3e