qcacmn: Fix stale vdev pointer access in case of ML LPC
Currently in case of ML LPC During Vdev attach mon_mac->mvdev
pointers are updated with same Vdev if mvdev is NULL as in case
of LPC single vdev is used for both MACs. But during vdev detach
only 1st mon_mac->mvdev pointer is set to NULL, Due to which when
Vdev attach is done again as 2nd mon_mac->mvdev is not NULL so
it's not updated to new vdev which results in 2nd mon_mac handle
having stale vdev pointer.

To fix the issue during vdev detach loop though all mon_mac and
if mvdev and current vdev is same reset mvdev to NULL.

CRs-Fixed: 3868091
Change-Id: I12ac2312ab2a2478fc60984b6a34107158cf67ed