qcacmn: Handle reason for VDEV repurpose cb notify
Disconnect from SB comes while the VDEV repurpose in pending
queue will not be deferred and will move the VDEV state to
disconnecting and serialize the disconnect command and waits
behind VDEV repurpose command. Now the order of disconnect
will be actual partner VDEV followed by actual assoc VDEV.

When VDEV repurpose command gets active in serialization, it
checks if the VDEV is still in connected state to bail out
if state changes due to some other disconnect and notifies
all the registered handler in the error handling as well.

In MLO manager notify callback, checks to distinguish the
reason for notification is not present and when such
notification happens on assoc VDEV, then it calls OSIF to
change the deflink pointer and also changes the mlo/link
VDEV flags and later the lVDEV repurpose status is sent
to FW but the deflink is not restored.

Once VDEV repurpose goes out of serialization the disconnect
request on partner link gets active, however as deflink is
pointing to this partner link, it notifies kernel about
disconnect and framework may initiate set mac address while
still the assoc VDEV is disconnecting. If VDEV detach gets
called from set MAC address before peer delete due to disconnect,
DP will not be able to clean the MLO peer attached to that VDEV
and potentially result in reference leaks.

Change-Id: I82b5e886982996e14e9953fbfe9090f887f4103e
CRs-Fixed: 3868050