qcacmn: Enable flush dbr logic for IPQ807x
Problem : It is observed that the router stops receiving CSI
with IoT devices connected to 5 GHz interface after powering on
for a while. After rebooting the router, it starts receiving
CSI again but with a delay and eventually stops getting CSI
data.

RCA: Tx completion event is not received at host for multiple
DBR events. Hence, the DBR events gets accumulated and
eventually the ring gets full, causing a drop in further DBR
events.

Fix: The logic to flush the DBR ring after a certain time period
is already available for IPQ6018 onwards chipsets. This change
is responsible in availing the flush DBR logic for IPQ807x

Change-Id: Ifed03e066ff11cc048eb97c9b878ef829b38bbe2
CRs-Fixed: 3793100