qcacmn: Update Max Channel Switch Time value
In the current implementation, the Max Channel Switch Time value for
non-DFS channel is configured as 1200 milliseconds. For a DFS channel,
it is 62 seconds plus 1200 milliseconds. For weather radar channel, the
value configured as 602 seconds plus 1200 milliseconds.
This configured MCST value is more than STA BMISS value.

Configured maximum MLDs and triggered CSA. The time required in the host
driver to,
	1) process the CSA completion event, plus
	2) send multi-vdev restart WMI command to FW, plus
	3) process the vdev restart response for all the vdevs of the
	   impacted link, plus
	4) send the vdev up command for all the vdevs of the impacted link
is ~750 milliseconds.

Hence, replace Host vdev restart time 1 second with 750 milliseconds.

Change-Id: Ib169019ec587de8607291350b4e0f3f16b3a92ec
CRs-Fixed: 3895493