qcacmn: Check whether phymode is eht when generating scan entry
In current implementation, when generating scan entry using
util_scan_gen_scan_entry() API, phymode is set before
updating ml info for that scan entry. If the phymode entry
is non-eht then the candidate is not ML capable, but if ML IEs
for that scan entry are present this creates a disperancy when
connection manager checks ML IEs for setting AP as ML capable
in cm_is_ml_connection() API.

To fix this, if phymode is non-eht and EHT cap or ML IEs are
present then set the ML IEs to NULL and reset EHT info when
generating scan entry in util_scan_gen_scan_entry(). Also check,
if phymode is eht in current candidate before setting the vdev
as ml capable in cm_is_ml_connection().

CRs-Fixed: 4010162
Change-Id: I239d287b098b1ea430ff49e305b20234ab045efd