qcacmn: Remove BDF board ID log print from wma service ready
Currently there is a provision to send the board ID by FW in
WMI service ready but this interface is no more used to send
this info and host is pritning this info as 0. This Board ID
information is sent using the QMI interface.

So remove the print in service ready indication for this
board ID.

Change-Id: I0f00d4d33ada41ec2d1df4b72d30b3c52ac607ea
CRs-Fixed: 4014892