qcacmn: Reset bss if link bss is not found
Host driver sends the same bss for both assoc and partner
links, leading to undesired behavior in kernel.

Scenario:
1) Host driver fills the assoc bss in link id 0 to handle
roamed indication for the legacy connections.
2) The assoc link bss is filled in link idx 1.
3) The patner link bss is not found, therefore not filled
in link idx 0.
4) The assoc bss filed in #1 to handle legacy connection
gets retained. Therefore, host ends up sending same bss for
two links.

To fix this, reset the bss entry on the link index of the
partner link in the roamed indication if the bss is not
found.

Change-Id: I9e941e4a9ff9760ada4c8c42015c925cc08b528a
CRs-Fixed: 4106746