qcacmn: Update policy mgr entry only if VDEV is connected
On MLO link state switch event from FW, host tries to update the
entries in policy manager table according to the new state of the
link. Processing this event while any VDEV is not in connected
state (disconnect sequence is on-going) can result in race condition
and potentially corrupt the policy manager entries.

Check the state of VDEV before updating the policy manager entry
to prevent possibility of stale entries due to parallel access.

Change-Id: I64abb5251e3a9a1b59a7b2315e9ed1e353f3a406
CRs-Fixed: 4179247