qcacmn: Define modified_in_curr_session with qdf_bitmap
Currently, the modified_in_curr_session is defined as an
unsigned long, and kernel APIs (set_bit, test_bit, clear_bit)
are used to manipulate it. However, these APIs might expect
an unsigned long array, and a static analyzer tool has
flagged this as an out-of-bounds (OOB) access.

To address this, use QDF APIs to define and operate on the
bitmap, ensuring proper handling.

Change-Id: Ie4a1ef55f046ca2bf481964e22be928b16ece7c9
CRs-Fixed: 4262079