qcacmn: ignore "no such process(listeners)" error
When CONFIG_CNSS_GENL is enabled and the cnss_diag
program is not running, a large number of logs are
continuously printed to the serial console, which
affects normal usage.
This occurs because sending a Netlink message, but
no processes (listeners) matching the filter to
receive the broadcast message. Therefore, this error
can be safely ignored.

Change-Id: I5ebbb02226d1297fda9b16e50e9c40be23fec889
CRs-Fixed: 4288270