qcacmn: Move the bufptr after populating current TLV in wmi buffer
Host driver populates the TLV corresponds to SA-query when
userspace sends it. But the buf ptr is not moved to the
end of TLV and it results in overwriting the filled TLV
with a new TLV(e.g. OFDMA).
Move the buf ptr to the end of the current filled TLV to allow
new TLVs without corrupting the original TLV.
Same is applicable for OFDMA TLV.

Change-Id: I88e440bd7208123ea0116929cd8ae1a512fae094
CRs-Fixed: 4337070